fix: corregir formulario y mover a carpeta components
